highlight connections (stage indicator removed)p
