--- client/src/components/StageIndicator/stageIndicator.css
@@ -1,85 +1,0 @@
-.stage-indicator {
-    display: flex;
-    align-items: center;
-    position: absolute;
-    top: 80px;
-    right: 10px;
-    background: rgba(255, 255, 255, 0.9);
-    border-radius: 16px;
-    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15);
-    backdrop-filter: blur(10px);
-    padding: 8px;
-    overflow: visible;
-    transition: width 0.4s cubic-bezier(0.25, 1, 0.5, 1), padding 0.3s ease;
-    cursor: pointer;
-    padding: 10px;
-}
-
-.stage-indicator.collapsed {
-    width: 40px; /* Only shows a single circle when collapsed */
-    padding: 8px 12px;
-    justify-content: center; /* Ensure the current stage is always centered */
-}
-
-.stage-indicator.expanded {
-    width: auto;
-    padding: 8px;
-}
-
-.stage-container {
-    display: flex;
-    align-items: center;
-    opacity: 1;
-    transition: opacity 0.3s ease, transform 0.3s cubic-bezier(0.25, 1, 0.5, 1);
-}
-
-.stage-container.hidden {
-    opacity: 0;
-    transform: scale(0);
-    pointer-events: none;
-}
-
-
-.circle {
-    width: 30px;
-    height: 30px;
-    line-height: 30px;
-    border-radius: 50%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    font-size: 14px;
-    color: #333;
-    background: linear-gradient(145deg, #e0e0e0, #f9f9f9);
-    cursor: pointer;
-    transition: all 0.4s cubic-bezier(0.25, 1, 0.5, 1);
-    box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.2);
-    user-select: none;
-    padding: 5px;
-    box-sizing: content-box;
-}
-
-
-.circle.active {
-    background: linear-gradient(145deg, #007aff, #559dff);
-    color: #fff;
-    transform: scale(1.15);
-    box-shadow: 0px 6px 15px rgba(0, 122, 255, 0.4);
-}
-
-.line {
-    width: 40px;
-    height: 4px;
-    border-radius: 2px;
-    background-color: transparent;
-    margin: 0 4px;
-    transition: width 0.3s ease, background-color 0.3s ease;
-}
-
-.stage-indicator.expanded .line {
-    background-color: #e0e0e0;
-}
-
-.line.active {
-    background: linear-gradient(145deg, #007aff, #559dff);
-}
